+++ lib/flutter_gyw.dart
@@ -2,9 +2,10 @@
 /// Get Your Way devices like the aRdent smart glasses.
 library flutter_gyw;
 
-export 'src/bt_device.dart' show BTDevice;
-export 'src/bt_manager.dart' show BTManager;
-export 'src/drawings.dart' show Drawing, TextDrawing, WhiteScreen, IconDrawing;
+export 'src/bt_device.dart' show GYWBtDevice;
+export 'src/bt_manager.dart' show GYWBtManager;
+export 'src/drawings.dart'
+    show GYWDrawing, TextDrawing, WhiteScreen, IconDrawing;
 export 'src/exceptions.dart' show GYWException, GYWStatusException;
 export 'src/fonts.dart' show GYWFont, GYWFonts;
 export 'src/icons.dart' show GYWIcon, GYWIcons;